Add better display for over/under, improve settings

+++ ui/servsettings.py
@@ -4,7 +4,6 @@ from contextlib import suppress
 from typing import List, Optional, TYPE_CHECKING, TypeVar
 
 import disnake
-import pydantic
 
 from utils.aldclient import discord_user_to_dict
 from utils.functions import natural_join
@@ -45,6 +44,15 @@ class ServerSettingsMenuBase(MenuBase, abc.ABC):
                 "and react with :game_die: - click the reaction to roll!"
             )
         return "Inline rolling is currently **enabled**. I'll roll any `[[dice]]` I find in messages!"
+
+    def get_over_under_desc(self, setting: str = None) -> str:
+        settings = getattr(self.settings, f"randchar_{setting}")
+        if not settings:
+            return "None"
+        out = []
+        for val, num in settings.items():
+            out.append(f"{num} {setting} {val}")
+        return f"At least {', '.join(out)}"
 
 
 class ServerSettingsUI(ServerSettingsMenuBase):
@@ -99,8 +107,8 @@ class ServerSettingsUI(ServerSettingsMenuBase):
             f"**Assign Stats**: {self.settings.randchar_straight}\n"
             f"**Minimum Total**: {self.settings.randchar_min}\n"
             f"**Maximum Total**: {self.settings.randchar_max}\n"
-            f"**Number over value**: {self.settings.randchar_over}\n"
-            f"**Number under value**: {self.settings.randchar_under}\n",
+            f"**Number over value**: {self.get_over_under_desc('over')}\n"
+            f"**Number under value**: {self.get_over_under_desc('under')}\n",
             inline=False,
         )
 
@@ -381,7 +389,7 @@ class _MiscellaneousSettingsUI(ServerSettingsMenuBase):
 
 class _RandcharSettingsUI(ServerSettingsMenuBase):
     # ==== ui ====
-    @disnake.ui.button(label="Select Dice", style=disnake.ButtonStyle.primary)
+    @disnake.ui.button(label="Set Dice", style=disnake.ButtonStyle.primary)
     async def select_dice(self, button: disnake.ui.Button, interaction: disnake.Interaction):
         button.disabled = True
         await self.refresh_content(interaction)
@@ -399,7 +407,7 @@ class _RandcharSettingsUI(ServerSettingsMenuBase):
             with suppress(disnake.HTTPException):
                 await input_msg.delete()
         except (ValueError, asyncio.TimeoutError):
-            await interaction.send("No valid dice found. Press `Select Dice` to try again.", ephemeral=True)
+            await interaction.send("No valid dice found. Press `Set Dice` to try again.", ephemeral=True)
         else:
             await self.commit_settings()
             await interaction.send("Your dice have been updated.", ephemeral=True)
@@ -407,7 +415,7 @@ class _RandcharSettingsUI(ServerSettingsMenuBase):
             button.disabled = False
             await self.refresh_content(interaction)
 
-    @disnake.ui.button(label="Select Number of Sets", style=disnake.ButtonStyle.primary)
+    @disnake.ui.button(label="Set Number of Sets", style=disnake.ButtonStyle.primary)
     async def select_sets(self, button: disnake.ui.Button, interaction: disnake.Interaction):
         button.disabled = True
         await self.refresh_content(interaction)
@@ -421,12 +429,14 @@ class _RandcharSettingsUI(ServerSettingsMenuBase):
                 timeout=60,
                 check=lambda msg: msg.author == interaction.author and msg.channel.id == interaction.channel_id,
             )
+            if not 1 <= int(input_msg.content) <= 25:
+                raise ValueError
             self.settings.randchar_sets = int(input_msg.content)
             with suppress(disnake.HTTPException):
                 await input_msg.delete()
         except (ValueError, asyncio.TimeoutError):
             await interaction.send(
-                "No valid number of sets found. Press `Select Number of Sets` to try again.", ephemeral=True
+                "No valid number of sets found. Press `Set Number of Sets` to try again.", ephemeral=True
             )
         else:
             await self.commit_settings()
@@ -441,7 +451,7 @@ class _RandcharSettingsUI(ServerSettingsMenuBase):
         await self.commit_settings()
         await self.refresh_content(interaction)
 
-    @disnake.ui.button(label="Select Minimum Score", style=disnake.ButtonStyle.primary)
+    @disnake.ui.button(label="Set Minimum", style=disnake.ButtonStyle.primary)
     async def select_minimum(self, button: disnake.ui.Button, interaction: disnake.Interaction):
         button.disabled = True
         await self.refresh_content(interaction)
@@ -459,7 +469,7 @@ class _RandcharSettingsUI(ServerSettingsMenuBase):
             with suppress(disnake.HTTPException):
                 await input_msg.delete()
         except (ValueError, asyncio.TimeoutError):
-            await interaction.send("No valid minimum found. Press `Select Minimum Score` to try again.", ephemeral=True)
+            await interaction.send("No valid minimum found. Press `Set Minimum` to try again.", ephemeral=True)
         else:
             await self.commit_settings()
             await interaction.send("Your minimum score has been updated.", ephemeral=True)
@@ -467,7 +477,7 @@ class _RandcharSettingsUI(ServerSettingsMenuBase):
             button.disabled = False
             await self.refresh_content(interaction)
 
-    @disnake.ui.button(label="Select Maximum Score", style=disnake.ButtonStyle.primary)
+    @disnake.ui.button(label="Set Maximum", style=disnake.ButtonStyle.primary)
     async def select_maximum(self, button: disnake.ui.Button, interaction: disnake.Interaction):
         button.disabled = True
         await self.refresh_content(interaction)
@@ -485,7 +495,7 @@ class _RandcharSettingsUI(ServerSettingsMenuBase):
             with suppress(disnake.HTTPException):
                 await input_msg.delete()
         except (ValueError, asyncio.TimeoutError):
-            await interaction.send("No valid maximum found. Press `Select Maximum Score` to try again.", ephemeral=True)
+            await interaction.send("No valid maximum found. Press `Set Maximum` to try again.", ephemeral=True)
         else:
             await self.commit_settings()
             await interaction.send("Your maximum score has been updated.", ephemeral=True)
@@ -493,12 +503,13 @@ class _RandcharSettingsUI(ServerSettingsMenuBase):
             button.disabled = False
             await self.refresh_content(interaction)
 
-    @disnake.ui.button(label="Add Required Over Stat", style=disnake.ButtonStyle.primary, row=1)
+    @disnake.ui.button(label="Add Required Over", style=disnake.ButtonStyle.primary, row=1)
     async def add_over(self, button: disnake.ui.Button, interaction: disnake.Interaction):
         button.disabled = True
         await self.refresh_content(interaction)
         await interaction.send(
-            "Choose a new minimum roll total by sending a message in this channel.",
+            "Choose a new required over by sending a message in this channel.\n"
+            "Please use the format 'score|number', for example '15|1' for at least one over 15",
             ephemeral=True,
         )
         try:
@@ -511,13 +522,16 @@ class _RandcharSettingsUI(ServerSettingsMenuBase):
             if self.settings.randchar_over is None:
                 self.settings.randchar_over = {}
             self.settings.randchar_over[value] = int(amount)
+            self._refresh_remove_over_select()
             with suppress(disnake.HTTPException):
                 await input_msg.delete()
         except (ValueError, asyncio.TimeoutError):
-            await interaction.send("No valid minimum found. Press `Select Minimum Score` to try again.", ephemeral=True)
+            await interaction.send(
+                "No valid required over found. Press `Add Required Over` to try again.", ephemeral=True
+            )
         else:
             await self.commit_settings()
-            await interaction.send("Your minimum score has been updated.", ephemeral=True)
+            await interaction.send("Your required over scores has been updated.", ephemeral=True)
         finally:
             button.disabled = False
             await self.refresh_content(interaction)
@@ -527,7 +541,8 @@ class _RandcharSettingsUI(ServerSettingsMenuBase):
         button.disabled = True
         await self.refresh_content(interaction)
         await interaction.send(
-            "Choose a new maximum roll total by sending a message in this channel.",
+            "Choose a new required under by sending a message in this channel.\n"
+            "Please use the format 'score|number', for example '10|1' for at least one under 10",
             ephemeral=True,
         )
         try:
@@ -540,18 +555,21 @@ class _RandcharSettingsUI(ServerSettingsMenuBase):
             if self.settings.randchar_under is None:
                 self.settings.randchar_under = {}
             self.settings.randchar_under[value] = int(amount)
+            self._refresh_remove_under_select()
             with suppress(disnake.HTTPException):
                 await input_msg.delete()
         except (ValueError, asyncio.TimeoutError):
-            await interaction.send("No valid maximum found. Press `Select Maximum Score` to try again.", ephemeral=True)
+            await interaction.send(
+                "No valid required over found. Press `Add Required Over` to try again.", ephemeral=True
+            )
         else:
             await self.commit_settings()
-            await interaction.send("Your maximum score has been updated.", ephemeral=True)
+            await interaction.send("Your required under scores has been updated.", ephemeral=True)
         finally:
             button.disabled = False
             await self.refresh_content(interaction)
 
-    @disnake.ui.select(placeholder="Remove Over", min_values=0, row=2)
+    @disnake.ui.select(placeholder="Remove Required Over", min_values=0, max_values=1, row=2)
     async def remove_over(self, select: disnake.ui.Select, interaction: disnake.Interaction):
         removed_over = select.values
         for remove in removed_over:
@@ -560,7 +578,7 @@ class _RandcharSettingsUI(ServerSettingsMenuBase):
         await self.commit_settings()
         await self.refresh_content(interaction)
 
-    @disnake.ui.select(placeholder="Remove Under", min_values=0, row=3)
+    @disnake.ui.select(placeholder="Remove Required Under", min_values=0, max_values=1, row=3)
     async def remove_under(self, select: disnake.ui.Select, interaction: disnake.Interaction):
         removed_under = select.values
         for remove in removed_under:
@@ -575,28 +593,26 @@ class _RandcharSettingsUI(ServerSettingsMenuBase):
 
     # ==== content ====
     def _refresh_remove_over_select(self):
-        """Update the options in the Remove Over select to reflect the currently selected values."""
+        """Update the options in the Remove Over select to reflect the currently available values."""
         self.remove_over.options.clear()
         if not self.settings.randchar_over:
-            self.remove_over.add_option(label="Empty", value="Test")
+            self.remove_over.add_option(label="Empty")
             self.remove_over.disabled = True
             return
         self.remove_over.disabled = False
-        for over in self.settings.randchar_over:
-            self.remove_over.add_option(label=over, value=str(over))
-        self.remove_over.max_values = len(self.remove_over.options)
+        for over, num in self.settings.randchar_over.items():
+            self.remove_over.add_option(label=f"{num} over {over}", value=str(over))
 
     def _refresh_remove_under_select(self):
-        """Update the options in the Remove Over select to reflect the currently selected values."""
+        """Update the options in the Remove Over select to reflect the currently available values."""
         self.remove_under.options.clear()
         if not self.settings.randchar_under:
-            self.remove_under.add_option(label="Empty", value="Test")
+            self.remove_under.add_option(label="Empty")
             self.remove_under.disabled = True
             return
         self.remove_under.disabled = False
-        for under in self.settings.randchar_under:
-            self.remove_under.add_option(label=under, value=str(under))
-        self.remove_under.max_values = len(self.remove_under.options)
+        for under, num in self.settings.randchar_under.items():
+            self.remove_under.add_option(label=f"{num} under {under}", value=str(under))
 
     async def _before_send(self):
         self._refresh_remove_over_select()
@@ -638,14 +654,14 @@ class _RandcharSettingsUI(ServerSettingsMenuBase):
         )
         embed.add_field(
             name="Required Stats Over Certain Value",
-            value=f"**{self.settings.randchar_over}**\n"
+            value=f"**{self.get_over_under_desc('over')}**\n"
             f"*This is a list of how many of the stats you require to be over a certain value, "
             f"such as having at least 1 stat over 17.*",
             inline=False,
         )
         embed.add_field(
             name="Required Stats Under Certain Value",
-            value=f"**{self.settings.randchar_under}**\n"
+            value=f"**{self.get_over_under_desc('under')}**\n"
             f"*This is a list of how many of the stats you require to be under a certain value, "
             f"such as having at least 1 stat under 10.*",
             inline=False,
